requirements.txt
gpiozero==2.0.1
lgpio==0.2.2.0
rpi-lgpio==0.5
RPi.GPIO==0.7.1
colorzero==2.0
peewee==3.17.3
websockets==12.0
